Keep last_recorded_len's arithmetic wrapping
Three sites on one counter went from the transpile's `wrapping_sub` /
`wrapping_add` to plain operators during the rewrite, and two of them
panic in a debug build. `test_registers`'
`Test_recording_with_select_mode` reaches both:

    vgetc:      last_recorded_len -= last_vgetc_recorded_len   (subtract
                with overflow -> SIGABRT)
    gotchars:   last_recorded_len += state.buflen              (add with
                overflow, once the value has wrapped)
    ungetchars: last_recorded_len -= len

Upstream's comment on the first one says `last_recorded_len` can be
*larger* than what the previous `vgetc` recorded; it can also be
smaller, because `ungetchars` shrinks it and nothing re-syncs the pair.
C's `size_t` then wraps, and the huge value makes `get_recorded`'s
`len >= last_recorded_len` fail, so nothing is trimmed off the
recording. That is the behaviour, not an accident to be corrected here,
so all three are wrapping again with the reason recorded on the static.

This is the standing `size_t` rule, and this slice is the reminder of
why it exists: a `wrapping_*` in a transpiled body is a fact about the
C, and the argument for replacing it has to be made per site, not per
file. Neither keysweep nor the twelve mapping/input functional specs
saw it -- only the oldtest did, and only through a Select-mode
recording.

keysweep IDENTICAL; test_registers 36/36; clippy still clean.

+++ crates/nvim/src/nvim/getchar.rs
@@ -257,7 +257,16 @@ static block_redo: GlobalCell<bool> = GlobalCell::new(false_0 != 0);
 static KeyNoremap: GlobalCell<::core::ffi::c_int> = GlobalCell::new(0 as ::core::ffi::c_int);
 static typebuf_init: GlobalCell<[uint8_t; 265]> = GlobalCell::new([0; 265]);
 static noremapbuf_init: GlobalCell<[uint8_t; 265]> = GlobalCell::new([0; 265]);
-static last_recorded_len: GlobalCell<size_t> = GlobalCell::new(0 as size_t);
+/// How many bytes the last `gotchars` recorded, so that `get_recorded` can
+/// drop the keys that stopped the recording.
+///
+/// Every arithmetic on this counter is **wrapping**, as the C's `size_t` is:
+/// `vgetc` subtracts what the previous call recorded and `ungetchars`
+/// subtracts what it took back, and either can take it below zero. A huge
+/// value then makes `get_recorded`'s `len >= last_recorded_len` fail and
+/// nothing is trimmed, which is what upstream does. `test_registers`'
+/// Test_recording_with_select_mode reaches it.
+static last_recorded_len: GlobalCell<size_t> = GlobalCell::new(0);
 static e_recursive_mapping: GlobalCell<[::core::ffi::c_char; 24]> = GlobalCell::new(unsafe {
     ::core::mem::transmute::<[u8; 24], [::core::ffi::c_char; 24]>(*b"E223: Recursive mapping\0")
 });
